commit
fix the bug in migrations

+++ database/migrations/2023_12_05_110356_create_sesis_table.php
@@ -12,11 +12,11 @@ return new class extends Migration
     public function up(): void
     {
         Schema::create('sesi', function (Blueprint $table) {
-            $table->integer('id_sesi');
+            $table->unsignedBigInteger('id_sesi');
             $table->primary('id_sesi');
             $table->unsignedBigInteger('user_id');
             $table->foreign('user_id')->references('no')->on('registers')->onDelete('cascade');
-            $table->unsignedBigInteger('id_kelas');
+            $table->unsignedInteger('id_kelas');
             $table->foreign('id_kelas')->references('id_kelas')->on('kelas')->onDelete('cascade');
             $table->integer('batas_jadwal');
             $table->timestamps();
